subtext interface changes
