+++ lib/request.js
@@ -486,13 +486,13 @@ internals.Request.prototype.generateResponse = function (source, options) {
 };
 
 
-internals.Peek = function (response) {
+internals.Peek = function (request) {
 
     Stream.Transform.call(this);
-    this._response = response;
+    this._request = request;
     this.once('finish', function () {
 
-        response.emit('finish');
+        request.emit('finish');
     });
 };
 
@@ -501,7 +501,7 @@ Hoek.inherits(internals.Peek, Stream.Transform);
 
 internals.Peek.prototype._transform = function (chunk, encoding, callback) {
 
-    this._response.emit('peek', chunk, encoding);
+    this._request.emit('peek', chunk, encoding);
     this.push(chunk, encoding);
     callback();
 };
