+++ lib/route.js
@@ -105,6 +105,7 @@ exports = module.exports = internals.Route = function (options, server, env) {
         this.settings.payload.maxBytes = this.settings.payload.maxBytes || this.server.settings.payload.maxBytes;
         this.settings.payload.uploads = this.settings.payload.uploads || this.server.settings.payload.uploads;
         this.settings.payload.failAction = this.settings.payload.failAction || 'error';
+        this.settings.payload.timeout = this.settings.payload.timeout !== undefined ? this.settings.payload.timeout : this.server.settings.timeout.client;
         if (this.settings.payload.allow) {
             this.settings.payload.allow = [].concat(this.settings.payload.allow);
         }
@@ -247,7 +248,10 @@ internals.payload = function (request, next) {
         return next();
     }
 
-    Subtext.parse(request, request.route.payload, function (err) {
+    Subtext.parse(request.raw.req, request._tap(), request.route.payload, function (err, parsed) {
+
+        request.mime = parsed.mime;
+        request.payload = parsed.payload || null;
 
         if (!err) {
             return next();
