+++ lib/schema.js
@@ -116,9 +116,9 @@ internals.serverBase = Joi.object({
         })
     }),
     timeout: Joi.object({
-        socket: Joi.number().positive().allow(false),
-        client: Joi.number().positive().allow(false).required(),
-        server: Joi.number().positive().allow(false).required()
+        socket: Joi.number().integer().positive().allow(false),
+        client: Joi.number().integer().positive().allow(false).required(),
+        server: Joi.number().integer().positive().allow(false).required()
     }),
     tls: Joi.object().allow(null),
     views: Joi.object(),                                                        // Validated in vision
@@ -228,7 +228,8 @@ internals.routeConfig = Joi.object({
         override: Joi.string(),
         maxBytes: Joi.number(),
         uploads: Joi.string(),
-        failAction: Joi.string().valid('error', 'log', 'ignore')
+        failAction: Joi.string().valid('error', 'log', 'ignore'),
+        timeout: Joi.number().integer().positive().allow(false)
     }),
     auth: internals.auth.allow(false),
     validate: Joi.object({
